new_requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
cssselect2==0.9.0
defusedxml==0.7.1
distro==1.9.0
fonttools==4.62.0
fpdf2==2.8.7
freetype-py==2.5.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.0
google-genai==1.66.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
Markdown==3.10.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.17.0
numpy==2.4.3
openai==2.26.0
orjson==3.11.7
oscrypto==1.3.0
packaging==24.2
pandas==2.3.3
pillow==12.1.1
pinecone==8.1.0
pinecone-plugin-assistant==3.0.2
pinecone-plugin-interface==0.0.7
plantuml==0.3.0
protobuf==6.33.5
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyHanko==0.34.1
pyhanko-certvalidator==0.30.1
PyMuPDF==1.27.1
pyparsing==3.3.2
pypdf==6.8.0
PyPDF2==3.0.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
streamlit==1.55.0
svglib==1.6.0
tenacity==9.1.4
tinycss2==1.5.1
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
watchdog==6.0.0
webencodings==0.5.1
websockets==16.0
wkhtmltopdf==0.2
